+++ tools/rosbag_storage/include/rosbag/encryptor.h
@@ -46,10 +46,6 @@
 
 #include <boost/function.hpp>
 
-#ifndef _WIN32
-  #include <gpgme.h>
-  #include <openssl/aes.h>
-#endif
 
 namespace rosbag {
 
@@ -133,72 +129,6 @@ public:
     virtual bool readEncryptedHeader(boost::function<bool(ros::Header&)> read_header, ros::Header& header, Buffer& header_buffer, ChunkedFile& file) = 0;
 };
 
-class NoEncryptor : public EncryptorBase
-{
-public:
-    NoEncryptor() { }
-    ~NoEncryptor() { }
-
-    void initialize(Bag const&, std::string const&) { }
-    uint32_t encryptChunk(const uint32_t, const uint64_t, ChunkedFile&);
-    void decryptChunk(ChunkHeader const&, Buffer&, ChunkedFile&) const;
-    void addFieldsToFileHeader(ros::M_string&) const { }
-    void readFieldsFromFileHeader(ros::M_string const&) { }
-    void writeEncryptedHeader(boost::function<void(ros::M_string const&)>, ros::M_string const&, ChunkedFile&);
-    bool readEncryptedHeader(boost::function<bool(ros::Header&)>, ros::Header&, Buffer&, ChunkedFile&);
-};
-
-#ifndef _WIN32
-//! Initialize GPGME library
-/*!
- * This method initializes GPGME library, and set locale.
- */
-void initGpgme();
-
-//! Get GPG key
-/*!
- * \param ctx GPGME context
- * \param user User name of the GPG key
- * \param key GPG key found
- *
- * This method outputs a GPG key in the system keyring corresponding to the given user name.
- * This method throws BagException if the key is not found or error occurred.
- */
-void getGpgKey(gpgme_ctx_t& ctx, std::string const& user, gpgme_key_t& key);
-
-class AesCbcEncryptor : public EncryptorBase
-{
-public:
-    static const std::string GPG_USER_FIELD_NAME;
-    static const std::string ENCRYPTED_KEY_FIELD_NAME;
-
-public:
-    AesCbcEncryptor() { }
-    ~AesCbcEncryptor() { }
-
-    void initialize(Bag const& bag, std::string const& gpg_key_user);
-    uint32_t encryptChunk(const uint32_t chunk_size, const uint64_t chunk_data_pos, ChunkedFile& file);
-    void decryptChunk(ChunkHeader const& chunk_header, Buffer& decrypted_chunk, ChunkedFile& file) const;
-    void addFieldsToFileHeader(ros::M_string& header_fields) const;
-    void readFieldsFromFileHeader(ros::M_string const& header_fields);
-    void writeEncryptedHeader(boost::function<void(ros::M_string const&)>, ros::M_string const& header_fields, ChunkedFile&);
-    bool readEncryptedHeader(boost::function<bool(ros::Header&)>, ros::Header& header, Buffer& header_buffer, ChunkedFile&);
-
-private:
-    void buildSymmetricKey();
-
-private:
-    // User name of GPG key used for symmetric key encryption
-    std::string gpg_key_user_;
-    // Symmetric key for encryption/decryption
-    std::basic_string<unsigned char> symmetric_key_;
-    // Encrypted symmetric key
-    std::string encrypted_symmetric_key_;
-    // AES keys for encryption/decryption
-    AES_KEY aes_encrypt_key_;
-    AES_KEY aes_decrypt_key_;
-};
-#endif
 }
 
 #endif
